% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statistic.R
\name{calculate_ptt}
\alias{calculate_ptt}
\title{function to calculate the ptt statistic for a tree}
\usage{
calculate_ptt(tree, dt = 1e-04)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{dt}{time step of the discrete integration}
}
\description{
this function calculates the ptt statistic for a tree
}
