% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_ML.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence in Parameters}
\usage{
check_convergence(current_means, previous_means, tolerance = 1e-04)
}
\arguments{
\item{tolerance}{Numeric value specifying the convergence tolerance.}

\item{current_params}{Numeric vector of current parameter estimates.}

\item{previous_params}{Numeric vector of previous parameter estimates.}
}
\value{
Logical; TRUE if the change in all parameters is less than or equal to the tolerance, FALSE otherwise.
}
\description{
Determines if the parameter estimates have stabilized (converged) based on a specified tolerance.
}
\examples{
current_params <- c(0.5, 0.5, 0.5, 0.5)
previous_params <- c(0.49, 0.51, 0.49, 0.51)
tolerance <- 0.05
has_converged <- check_convergence_in_parameters(current_params, previous_params, tolerance)
print(has_converged)  # Should return TRUE if the change in parameters is within the tolerance

}
