% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_GAM.R
\name{dd_ML_est}
\alias{dd_ML_est}
\title{Optimize GAM Model Parameters with Fixed betaP}
\usage{
dd_ML_est(pgam)
}
\arguments{
\item{pgam}{A GAM model object for which the log likelihood is to be optimized. 
The model should include mu, lambda, betaN, and betaP as parameters.}
}
\value{
A list containing the optimized parameter values (pars) and the minimized 
        log likelihood value (ml).
}
\description{
This function optimizes the parameters of a given Generalized Additive Model (GAM) 
with the constraint that betaP is fixed at zero. It applies penalties to discourage 
selecting parameter values near the boundaries of their respective intervals, aiming 
to find optimal values of mu, lambda, and betaN that minimize the estimated log likelihood.
}
\examples{
# Assuming you have a GAM model 'pgam' fitted with parameters including mu, lambda, betaN, betaP
result <- dd_ML_est(pgam)
print(result$pars)  # Optimized parameters
print(result$ml)    # Minimized log likelihood

}
