% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.eval_particles}
\alias{.eval_particles}
\title{Dispatch particle evaluation to the appropriate mode}
\usage{
.eval_particles(pop, input, num_threads)
}
\description{
Dispatch particle evaluation to the appropriate mode
}
\keyword{internal}
