% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_cpp}
\alias{em_cpp}
\title{function to perform one step of the E-M algorithm}
\usage{
em_cpp(
  brts,
  init_pars,
  sample_size,
  maxN,
  max_missing,
  max_lambda,
  lower_bound,
  upper_bound,
  xtol_rel,
  num_threads,
  copy_trees,
  model = as.integer(c(0, 0, 0)),
  link = 0L,
  rho = 1,
  rconditional = NULL,
  parent_tip_start = as.numeric(c())
)
}
\arguments{
\item{brts}{vector of branching times}

\item{init_pars}{vector of initial parameter files}

\item{sample_size}{number of samples}

\item{maxN}{maximum number of failed trees}

\item{max_missing}{maximum number of species missing}

\item{max_lambda}{maximum speciation rate}

\item{lower_bound}{vector of lower bound values for optimization, should
be equal in length to the vector of init_pars}

\item{upper_bound}{vector of upper bound values for optimization, should
be equal in length to the vector of init_pars}

\item{xtol_rel}{relative tolerance for optimization}

\item{num_threads}{number of threads used.}

\item{copy_trees}{if set to true, the trees generated are returned as well}

\item{model}{integer vector of length 3: c(use_N, use_P, use_E)}

\item{link}{link function: 0 = linear (max(0,...)), 1 = exponential}

\item{rho}{sampling fraction in (0, 1]}

\item{rconditional}{R function that evaluates the GAM function.}

\item{parent_tip_start}{tip start of the lineage that splits at each
observed branching event, in the same forward-time order as \code{brts};
empty (the default) when only branching times are available, in which case
every observed lineage is recorded as dating from the crown and D = 0 at
every observed event.}
}
\value{
a list with the following components:
\describe{
 \item{trees}{list of trees}
 \item{rejected}{number of rejected trees}
 \item{rejected_overruns}{number of trees rejected due to too large size}
 \item{rejected_lambda}{number of trees rejected due to lambda errors}
 \item{rejected_zero_weights}{number of trees rejected due to zero weight (log weight -Inf)}
 \item{rejected_nonfinite}{number of trees rejected due to a +Inf or NaN log weight}
 \item{num_trees}{number of trees the E-step returned}
 \item{estimates}{vector of estimates}
 \item{nlopt}{nlopt status}
 \item{fhat}{vector of fhat values}
 \item{time}{time elapsed}
 \item{weights}{vector of weights}
 \item{logf}{vector of log p(obs, z_i | theta) for each valid tree}
 \item{logg}{vector of log q(z_i | obs, theta) for each valid tree}
}
}
\description{
function to perform one step of the E-M algorithm
}
\keyword{internal}
