% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{emphasis_de_factorial}
\alias{emphasis_de_factorial}
\title{perform emphasis analysis using DE method}
\usage{
emphasis_de_factorial(
  brts,
  num_iterations,
  num_points,
  max_missing,
  sd_vec,
  lower_bound,
  upper_bound,
  maxN = 10,
  max_lambda,
  disc_prop = 0.5,
  verbose = FALSE,
  num_threads = 1,
  model = c(0L, 0L, 0L),
  link = 0L
)
}
\arguments{
\item{brts}{branching times of tree to fit on}

\item{num_iterations}{number of iterations of the DE algorithm}

\item{num_points}{number of particles per iteration}

\item{max_missing}{maximum number of missing trees}

\item{sd_vec}{vector of initial values of standard deviation for perturbation}

\item{lower_bound}{vector of lower bound values for parameters,
used to populate the particles}

\item{upper_bound}{vector of upper bound values for parameters, used to
populate the particles}

\item{maxN}{maximum number of tries per parameter combination before giving
up}

\item{max_lambda}{maximum value of lambda}

\item{disc_prop}{proportion of particles retained per iteration}

\item{verbose}{verbose output if TRUE}

\item{num_threads}{number of threads}

\item{model}{integer vector of length 3: c(use_N, use_P, use_E)}

\item{link}{link function: 0 = linear, 1 = exponential}
}
\description{
perform emphasis analysis using DE method
}
