% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mcem.R
\name{estimation_mcem}
\alias{estimation_mcem}
\title{Monte Carlo EM Estimation for emphasis Package}
\usage{
estimation_mcem(phylo, model, ...)
}
\arguments{
\item{phylo}{A phylo object.}

\item{model}{Model specification list.}

\item{...}{Additional parameters for the estimation.}
}
\value{
A list with parameter estimates and diagnostics.
}
\description{
Fits a diversification model to a phylogenetic tree using the Monte Carlo Expectation-Maximization algorithm.
}
\details{
This file contains the MCEM estimation method for phylogenetic models.
}
\section{Functions}{

- estimation_mcem: Fit a diversification model using MCEM.
Monte Carlo EM Estimation
}

\examples{
# Example usage (pseudo-code):
# result <- estimation_mcem(phylo, model, ...)
}
