% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle.R
\name{estimation_mle}
\alias{estimation_mle}
\title{Maximum Likelihood Estimation for emphasis Package}
\usage{
estimation_mle(phylo, model, ...)
}
\arguments{
\item{phylo}{A phylo object.}

\item{model}{Model specification list.}

\item{...}{Additional parameters for the estimation.}
}
\value{
A list with parameter estimates and diagnostics.
}
\description{
Fits a diversification model to a phylogenetic tree using Maximum Likelihood Estimation.
}
\details{
This file contains the MLE estimation method for phylogenetic models.
}
\section{Functions}{

- estimation_mle: Fit a diversification model using MLE.
Maximum Likelihood Estimation
}

\examples{
# Example usage (pseudo-code):
# result <- estimation_mle(phylo, model, ...)
}
