% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_utility}
\alias{example_utility}
\title{Utility Functions for emphasis Package}
\usage{
example_utility(x)
}
\description{
This file contains internal utility functions for the emphasis package.
Example Utility Function
}
\keyword{internal}
