% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditional_likelihood.R
\name{fit_gam_survival}
\alias{fit_gam_survival}
\title{fit gam survival}
\usage{
fit_gam_survival(simulations, splines = "bivariate")
}
\arguments{
\item{simulations}{a dataframe containing the simulations}

\item{splines}{either bivariate or univariate}
}
\description{
function to fit a GAM function to simulated extinction data.
}
