% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_de.R
\name{learn_de}
\alias{learn_de}
\title{perform emphasis analysis using DE method}
\usage{
learn_de(
  phylo,
  max_iterations = 100,
  num_points = 1000,
  max_missing = 1e+05,
  lower_bound = c(0, 0, -0.1, -0.1),
  upper_bound = c(1, 5, 0.1, 0.1),
  verbose = TRUE,
  num_threads = 1
)
}
\arguments{
\item{max_iterations}{number of iterations of the DE algorithm}

\item{num_points}{number of particles per iteration}

\item{max_missing}{maximum number of missing trees}

\item{lower_bound}{vector of lower bound values for parameters,
used to populate the particles}

\item{upper_bound}{vector of upper bound values for parameters, used to
populate the particles}

\item{verbose}{verbose output if TRUE}

\item{num_threads}{number of threads}

\item{brts}{branching times of tree to fit on}

\item{maxN}{maximum number of tries per parameter combination before giving
up}

\item{max_lambda}{maximum value of lambda}

\item{disc_prop}{proportion of particles retained per iteration}
}
\description{
perform emphasis analysis using DE method
}
