% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_plugin.R
\name{locate_plugin}
\alias{locate_plugin}
\title{function to locate a diversification model}
\usage{
locate_plugin(PlugIn)
}
\arguments{
\item{PlugIn}{the name of the plugin}
}
\value{
path of the plugin.
}
\description{
function to locate a diversification model
}
