% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_cpp}
\alias{m_cpp}
\title{function to perform one step of the E-M algorithm}
\usage{
m_cpp(
  e_step,
  init_pars,
  plugin,
  lower_bound,
  upper_bound,
  xtol_rel,
  num_threads,
  model = as.integer(c(0, 0, 0)),
  link = 0L,
  rho = 1,
  rconditional = NULL
)
}
\arguments{
\item{e_step}{result of e_step function, as a list}

\item{init_pars}{vector of initial parameter values}

\item{plugin}{string indicating plugin used, currently available: 'rpd1' and
'rpd5c'}

\item{lower_bound}{vector of lower bound values for optimization, should
be equal in length to the vector of init_pars}

\item{upper_bound}{vector of upper bound values for optimization, should
be equal in length to the vector of init_pars}

\item{xtol_rel}{relative tolerance for optimization}

\item{num_threads}{number of threads used.}

\item{model}{integer vector of length 3: c(use_N, use_P, use_E)}

\item{link}{link function: 0 = linear (max(0,...)), 1 = exponential}

\item{rho}{Sampling fraction (0, 1]. Default \code{1} (complete sampling).}

\item{rconditional}{R function that evaluates the GAM function.}
}
\value{
list with the following entries:
\describe{
 \item{estimates}{vector of estimates}
 \item{nlopt}{nlopt status}
 \item{time}{used computation time}
}
}
\description{
function to perform one step of the E-M algorithm
}
\keyword{internal}
