% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.cem_diagnostics}
\alias{print.cem_diagnostics}
\title{Print method for CEM diagnostics}
\usage{
\method{print}{cem_diagnostics}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{diagnose_cem}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for CEM diagnostics
}
