% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.emphasis_fit}
\alias{print.emphasis_fit}
\title{Print method for emphasis_fit objects}
\usage{
\method{print}{emphasis_fit}(x, ...)
}
\arguments{
\item{x}{An \code{emphasis_fit} object returned by \code{\link{estimate_rates}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for emphasis_fit objects
}
