% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.emphasis_fits}
\alias{print.emphasis_fits}
\title{Print method for the fits of \code{estimate_rates(model = "all")}}
\usage{
\method{print}{emphasis_fits}(x, ...)
}
\arguments{
\item{x}{An \code{emphasis_fits} object.}

\item{...}{Ignored.}
}
\description{
Print method for the fits of \code{estimate_rates(model = "all")}
}
