% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.emphasis_pipeline}
\alias{print.emphasis_pipeline}
\title{Print method for emphasis_pipeline objects}
\usage{
\method{print}{emphasis_pipeline}(x, ...)
}
\arguments{
\item{x}{An \code{emphasis_pipeline} result.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for emphasis_pipeline objects
}
