% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.gam_diagnostics}
\alias{print.gam_diagnostics}
\title{Print method for GAM diagnostics}
\usage{
\method{print}{gam_diagnostics}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{diagnose_gam}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for GAM diagnostics
}
