% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.mcem_diagnostics}
\alias{print.mcem_diagnostics}
\title{Print method for MCEM diagnostics}
\usage{
\method{print}{mcem_diagnostics}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{diagnose_mcem}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for MCEM diagnostics
}
