% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.model_selection}
\alias{print.model_selection}
\title{Print method for model_selection objects}
\usage{
\method{print}{model_selection}(x, ...)
}
\arguments{
\item{x}{A \code{model_selection} object from
\code{\link{select_diversification_model}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for model_selection objects
}
