% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rescale_tree}
\alias{rescale_tree}
\title{function to rescale the branching times of a phylogenetic tree}
\usage{
rescale_tree(phy, new_crown_age = 1)
}
\arguments{
\item{phy}{phy object}

\item{new_crown_age}{new crown age of the phylogenetic tree, set to 1 by 
default}
}
\value{
rescaled phylogeny as a phy object
}
\description{
function to rescale the branching times of a phylogenetic tree
}
