% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_sim.R
\name{sim_single_tree_pd_cpp}
\alias{sim_single_tree_pd_cpp}
\title{simulation function to simulate a tree under the pd model}
\usage{
sim_single_tree_pd_cpp(pars, max_t, max_lin = 1e+06, max_tries = 100)
}
\arguments{
\item{pars}{parameter vector with c(mu, lambda_0, beta_N, beta_P)}

\item{max_t}{crown age}

\item{max_lin}{number of lineages past which non-extinction is assumed.}

\item{max_tries}{maximum number of tries to get a non-extinct tree.}
}
\value{
list with: 1) tes - reconstructed tree, 2) tas - tree with extinct
lineages, 3) L = Ltable and 4) brts - branching times of the reconstructed
tree
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
