% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sim.R
\name{sim_tree_dd_minimal}
\alias{sim_tree_dd_minimal}
\title{simulation function to simulate a minimal dd model}
\usage{
sim_tree_dd_minimal(pars, max_t)
}
\arguments{
\item{pars}{parameter vector with c(mu, lambda_0, beta_N)}

\item{max_t}{crown age}
}
\value{
a vector with population size and time of extinction
(equal to the crown age in the absence of extinction)
}
\description{
super fast function to simulate the process of diversification
with diversity dependence
}
