% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_sim.R
\name{sim_tree_pd_grid}
\alias{sim_tree_pd_grid}
\title{simulation function to simulate a tree under the pd model}
\usage{
sim_tree_pd_grid(mu_vec, lambda_vec, b_n_vec, b_p_vec, max_t, num_repl)
}
\arguments{
\item{max_t}{crown age}

\item{num_repl}{number of replicates}

\item{pars}{parameter vector with c(mu, lambda_0, beta_N, beta_P)}
}
\value{
a list with the phylogeny, and a vector with population size, 
time of extinction (equal to the crown #' age in the absence of extinction), 
and the phylogenetic diversity at the time of extinction (or crown age).
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
