% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_sim.R
\name{sim_tree_pd_R}
\alias{sim_tree_pd_R}
\title{simulation function to simulate a tree under the pd model, returns the full
tree.}
\usage{
sim_tree_pd_R(pars, max_t)
}
\arguments{
\item{pars}{parameter vector with c(mu, lambda_0, beta_N, beta_P)}

\item{max_t}{crown age}
}
\value{
a list with the phylogeny, and a vector with population size, 
time of extinction (equal to the crown #' age in the absence of extinction), 
and the phylogenetic diversity at the time of extinction (or crown age).
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
