% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augmentPD}
\alias{augmentPD}
\title{Augment Phylogenetic Diversity using MCEM}
\usage{
augmentPD(
  phylo,
  pars,
  maxN,
  max_missing,
  lower_bound,
  upper_bound,
  num_threads = 1,
  sample_size = 1,
  soc = 2,
  xtol_rel = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{phylo}{A phylogenetic tree of class 'phylo'.}

\item{pars}{A numeric vector of initial parameters for the MCEM algorithm.}

\item{maxN}{Maximum number of augmentation iterations.}

\item{max_missing}{Maximum number of missing lineages allowed.}

\item{lower_bound}{Lower bound vector for parameter optimization.}

\item{upper_bound}{Upper bound vector for parameter optimization.}

\item{num_threads}{Number of parallel threads. Defaults to 1.}

\item{sample_size}{Sample size for MCEM. Defaults to 1.}

\item{soc}{Number of lineages at crown (2) or stem (1). Defaults to 2.}

\item{xtol_rel}{Relative convergence tolerance. Defaults to 1e-5.}

\item{verbose}{If TRUE, print progress messages. Defaults to FALSE.}
}
\value{
A list with augmented tree samples, importance weights, and log-likelihood estimate.
}
\description{
Augments a phylogenetic tree using Monte Carlo Expectation-Maximization
(MCEM), sampling missing lineages and estimating model parameters.
}
\examples{
\dontrun{
  library(ape)
  tree <- rphylo(10, 0.5, 0)
  result <- augmentPD(
    phylo = tree,
    pars = c(0.1, 0.5, -0.01, 0.01),
    maxN = 100,
    max_missing = 1000,
    lower_bound = c(-0.1, 0, -0.1, -0.1),
    upper_bound = c(0.5, 2, 0.1, 0.1)
  )
}
}
