% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi.R
\name{.augment_tree_bdi}
\alias{.augment_tree_bdi}
\title{BDI augmentation: draw sample_size augmented trees.}
\usage{
.augment_tree_bdi(
  tree,
  pars,
  model_bin = c(0L, 0L, 0L),
  sample_size = 1L,
  max_missing = 10000,
  link = 0L,
  rho = 1,
  use_gaussian_closure = TRUE
)
}
\description{
Returns same structure as augment_trees() C++ function:
list(trees, logf, logg, weights, fhat)
}
\keyword{internal}
