% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi.R
\name{.bdi_to_tree_df}
\alias{.bdi_to_tree_df}
\title{Convert BDI species list to the tree data frame format used by emphasis.
Columns: brts, n, t_ext, pd, tip_start, id, parent_id}
\usage{
.bdi_to_tree_df(species, bt, tp)
}
\description{
Convert BDI species list to the tree data frame format used by emphasis.
Columns: brts, n, t_ext, pd, tip_start, id, parent_id
}
\keyword{internal}
