% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_cpp}
\alias{e_cpp}
\title{function to perform one step of the E-M algorithm}
\usage{
e_cpp(
  brts,
  init_pars,
  sample_size,
  maxN,
  plugin,
  soc,
  max_missing,
  max_lambda,
  lower_bound,
  upper_bound,
  xtol_rel,
  num_threads
)
}
\description{
function to perform one step of the E-M algorithm
}
