% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_cpp}
\alias{em_cpp}
\title{function to perform one step of the E-M algorithm}
\usage{
em_cpp(
  brts,
  init_pars,
  sample_size,
  maxN,
  plugin,
  soc,
  max_missing,
  max_lambda,
  lower_bound,
  upper_bound,
  xtol_rel,
  num_threads,
  copy_trees,
  rconditional = NULL
)
}
\description{
function to perform one step of the E-M algorithm
}
