% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasis.R
\name{emphasis}
\alias{emphasis}
\title{Emphasis main function}
\usage{
emphasis(
  brts,
  init_par,
  soc = 2,
  model,
  lower_bound = numeric(0),
  upper_bound = numeric(0),
  max_lambda = 500,
  xtol = 0.001,
  em_tol = 0.25,
  sample_size_tol = 0.005,
  verbose = FALSE,
  return_trees = FALSE,
  max_missing = 10000,
  burnin_sample_size = 200,
  pilot_sample_size = seq(100, 1000, by = 100),
  burnin_iterations = 20,
  num_threads = 0,
  conditional = NULL
)
}
\arguments{
\item{brts}{vector of branching times of the tree for which the model has to
be fitted}

\item{init_par}{initial parameter values of the model}

\item{soc}{number of species at the root (1) or crown (2). Default is 2.}

\item{model}{model to be used}

\item{lower_bound}{vector of the lower limit of parameter values used by the
model. Set to -Infinity if left empty.}

\item{upper_bound}{vector of the upper limit of parameter values used by the
model. Set to +Infinity if left empty}

\item{max_lambda}{maximum speciation rate, default is 500. Should not be set 
too high to avoid extremely long run times}

\item{xtol}{tolerance of step size in the M step}

\item{em_tol}{tolerance of step size in cycling through EM}

\item{sample_size_tol}{tolerance in determining the sample size}

\item{verbose}{if TRUE, provides textual output of intermediate steps}

\item{return_trees}{boolean, if TRUE the simulated trees are returned as well}

\item{max_missing}{maximum number of tips a tree can be augmented with.}

\item{burnin_sample_size}{sample size during burn-in}

\item{pilot_sample_size}{vector of sample sizes used to determine the true 
sampling size}

\item{burnin_iterations}{number of iterations of the EM algorithm to discard
as burn-in}

\item{num_threads}{number of threads to be used. If set to 0, the maximum 
number of threads available is chosen.}

\item{conditional}{a function that takes a parameter set as argument and returns
conditional probability.}
}
\value{
a list with two components: 1) \code{pars} contains the average parameter
estimate and 2) \code{MCEM} matrix of parameter estimates and likelihoods.
}
\description{
Main function of emphasis, that uses an E-M approach to fit a
diversification model to a phylogenetic tree.
}
