% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasis_de.R
\name{emphasis_de}
\alias{emphasis_de}
\title{perform emphasis analysis using DE method}
\usage{
emphasis_de(
  brts,
  num_iterations,
  num_points,
  max_missing,
  sd_vec,
  lower_bound,
  upper_bound,
  max_lambda,
  disc_prop = 0.5,
  verbose = FALSE,
  num_threads = 1
)
}
\arguments{
\item{brts}{branching times of tree to fit on}

\item{num_iterations}{number of iterations of the DE algorithm}

\item{num_points}{number of particles per iteration}

\item{max_missing}{maximum number of missing trees}

\item{sd_vec}{vector of initial values of standard deviation for perturbation}

\item{lower_bound}{vector of lower bound values for parameters,
used to populate the particles}

\item{upper_bound}{vector of upper bound values for parameters, used to
populate the particles}

\item{max_lambda}{maximum value of lambda}

\item{disc_prop}{proportion of particles retained per iteration}

\item{verbose}{verbose output if TRUE}

\item{num_threads}{number of threads}
}
\description{
perform emphasis analysis using DE method
}
