% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNonHomogeneousExp.R
\name{generateNonHomogeneousExp}
\alias{generateNonHomogeneousExp}
\title{Generate Variates for a Non-Homogeneous Exponential Process}
\usage{
generateNonHomogeneousExp(num_variates, rate_func, start_time, max_time)
}
\arguments{
\item{num_variates}{The number of variates to generate.}

\item{rate_func}{A function representing the rate of the non-homogeneous process.}

\item{start_time}{The start time for the generation of variates.}

\item{max_time}{The maximum time limit for the generation of variates.}
}
\value{
A numeric vector of generated variates.
}
\description{
This function generates random variates for a non-homogeneous exponential
distribution using a thinning algorithm.
}
\examples{
rate_function_example <- function(t) { 2 * t }
variates <- generateNonHomogeneousExp(5, rate_function_example, 0, 10)
print(variates)
}
