% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNonHomogeneousExp.R
\name{generateNonHomogeneousExp}
\alias{generateNonHomogeneousExp}
\title{Generate Variates for a Non-Homogeneous Exponential Process}
\usage{
generateNonHomogeneousExp(
  num_variates,
  covariates,
  parameters,
  start_time,
  max_time
)
}
\arguments{
\item{num_variates}{The number of variates to generate.}

\item{covariates}{A numeric matrix of covariates where each row represents a 
set of covariates and the last column is expected to be 
time-dependent.}

\item{parameters}{Parameters for the exponential rate function, including the bias.}

\item{start_time}{The start time for the generation of variates.}

\item{max_time}{The maximum time limit for the generation of variates.}
}
\value{
A numeric vector of generated variates.

print(generated_variates)
}
\description{
This function generates random variates for a non-homogeneous exponential
distribution using a thinning algorithm.
}
