% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generatePhyloPD}
\alias{generatePhyloPD}
\title{Generate Phylogenetic Diversity Data}
\usage{
generatePhyloPD(
  n_trees,
  mu_interval,
  lambda_interval,
  betaN_interval,
  betaP_interval,
  max_lin = 1e+06,
  max_tries = 1
)
}
\arguments{
\item{n_trees}{Number of trees to simulate.}

\item{mu_interval}{Interval for mu parameter.}

\item{lambda_interval}{Interval for lambda parameter.}

\item{betaN_interval}{Interval for betaN parameter.}

\item{betaP_interval}{Interval for betaP parameter.}

\item{max_lin}{Maximum number of lineages.}

\item{max_tries}{Maximum number of tries for simulation.}
}
\value{
A list containing the simulated trees and parameters.
}
\description{
This function simulates phylogenetic trees based on given parameters.
}
\examples{
\donttest{
generatePhyloPD(n_trees = 5,
               mu_interval = c(0.05, 0.1),
               lambda_interval = c(0.3, 0.5),
               betaN_interval = c(-0.05, -0.01),
               betaP_interval = c(-0.01, 0.01),
               max_lin = 1e4,
               max_tries = 10)
}
}
