% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePhyloPD.R,
%   R/simulation_generatePhyloPD.R
\name{generatePhyloPD}
\alias{generatePhyloPD}
\title{Generate Phylogenetic trees under PD model}
\usage{
generatePhyloPD(
  n_trees,
  mu_interval,
  lambda_interval,
  betaN_interval,
  betaP_interval,
  max_lin = 1e+06,
  max_tries = 1
)

generatePhyloPD(
  n_trees,
  mu_interval,
  lambda_interval,
  betaN_interval,
  betaP_interval,
  max_lin = 1e+06,
  max_tries = 1
)
}
\arguments{
\item{n_trees}{Number of trees to simulate.}

\item{mu_interval}{Interval for mu parameter.}

\item{lambda_interval}{Interval for lambda parameter.}

\item{betaN_interval}{Interval for betaN parameter.}

\item{betaP_interval}{Interval for betaP parameter.}

\item{max_lin}{Maximum number of lineages.}

\item{max_tries}{Maximum number of tries for simulation.}
}
\value{
A list containing the simulated trees and parameters.

A list containing the simulated trees and parameters.
}
\description{
This function simulates phylogenetic trees based on given parameters.

This function simulates phylogenetic trees based on given parameters.
}
\examples{
generatePhyloPD(n_trees <- 10000,
               mu_interval <- c(0, 0.5),
               lambda_interval <- c(0.5, 3),
               betaN_interval <- c(-0.02, 0.02) ,
               betaP_interval <- c(-0.02, 0.02) )
               
generatePhyloPD(n_trees = 10, 
               mu_interval = c(0.1, 0.5), 
               lambda_interval = c(0.1, 0.5),
               betaN_interval = c(0.1, 0.5), 
               betaP_interval = c(0.1, 0.5),
               max_lin = 1e6,
               max_tries = 1)
}
