% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihood}
\alias{loglikelihood}
\title{function to calculate log likelihood of pars for a tree set,}
\usage{
loglikelihood(pars, trees, logg, plugin, num_rejected)
}
\description{
function to calculate log likelihood of pars for a tree set,
}
