% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_cpp}
\alias{m_cpp}
\title{function to perform one step of the E-M algorithm}
\usage{
m_cpp(
  e_step,
  init_pars,
  plugin,
  lower_bound,
  upper_bound,
  xtol_rel,
  num_threads,
  rconditional = NULL
)
}
\description{
function to perform one step of the E-M algorithm
}
