% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNonHomogeneousExp.R
\name{nhExpRand}
\alias{nhExpRand}
\title{Generate Non-Homogeneous Exponential Random Variables}
\usage{
nhExpRand(n, rate_func, now = 0, tMax = Inf)
}
\arguments{
\item{n}{The number of variables to generate.}

\item{rate_func}{The rate function of the non-homogeneous process.}

\item{now}{The current time, default is 0.}

\item{tMax}{The maximum time, default is Inf.}
}
\value{
A numeric vector of generated variables.
}
\description{
This function generates random variables from a non-homogeneous exponential
distribution based on a provided rate function.
}
\examples{
rate_function_example <- function(t) { 1 + sin(t) }
nh_exp_vars <- nhExpRand(10, rate_function_example, now = 0, tMax = 20)
print(nh_exp_vars)
}
