% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tree_pd.R
\name{sim_tree_pd_cpp}
\alias{sim_tree_pd_cpp}
\title{simulation function to simulate a tree under the pd model}
\usage{
sim_tree_pd_cpp(pars, max_t, max_lin = 1e+06, max_tries = 100, useDDD = TRUE)
}
\arguments{
\item{pars}{parameter vector with c(mu, lambda_0, beta_N, beta_P)}

\item{max_t}{crown age}

\item{max_lin}{number of lineages past which non-extinction is assumed.}

\item{max_tries}{maximum number of tries to get a non-extinct tree.}
}
\value{
list with: 1) tes - reconstructed tree, 2) tas - tree with extinct
lineages, 3) L = Ltable and 4) brts - branching times of the reconstructed
tree and 5) status of simulation, options 1) "extinct", 2) "too_large" or 3)
"done".
}
\example{
sim.param = c(0.1, 0.8, -0.01, 0.01)


tr = sim_tree_pd_cpp(pars = sim.param, 
                max_t = 30, 
                max_lin = 100000, 
                max_tries = max_tries,useDDD = TRUE)


plot(tr$tes)
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
