% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_sim.R
\name{sim_tree_pd_cpp}
\alias{sim_tree_pd_cpp}
\title{simulation function to simulate a tree under the pd model, returning whether
the tree went extinct before max_t or not. 
This function does not return a phylogenetic tree to improve computation speed.}
\usage{
sim_tree_pd_cpp(pars, max_t, num_repl = 1, max_N)
}
\arguments{
\item{pars}{parameter vector with c(mu, lambda_0, beta_N, beta_P)}

\item{max_t}{crown age}

\item{num_repl}{number of replicates}

\item{max_N}{number of lineages past which non-extinction is assumed.}
}
\value{
a tibble with four columns: 1) whether or not the tree went extinct,
2) the time of extinction (equal to the crown #' age in the absence of 
extinction), 3) the number of tips, and 4) the phylogenetic diversity at the 
time of extinction (or crown age).
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
