% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tree_pd.R
\name{sim_tree_pd_grid}
\alias{sim_tree_pd_grid}
\title{simulation function to simulate a tree under the pd model}
\usage{
sim_tree_pd_grid(mu_vec, lambda_vec, b_n_vec, b_p_vec, max_t, num_repl, max_N)
}
\arguments{
\item{mu_vec}{vector of extinction rates to explore}

\item{lambda_vec}{vector of lambda rates to explore}

\item{b_n_vec}{vector of B_n rates to explore}

\item{b_p_vec}{vector of B_p rates to explore}

\item{max_t}{crown age}

\item{num_repl}{number of replicates}

\item{max_N}{maximum number of lineages past which non-extinction is assumed.}
}
\value{
a list with the used parameter combinations, 
extinction of the phylogeny, and a vector with population size, 
time of extinction (equal to the crown #' age in the absence of extinction), 
and the phylogenetic diversity at the time of extinction (or crown age).
}
\description{
super fast function to simulate the process of diversification
with diversity dependence and phylogenetic diversity dependence
}
